requirements.txt
Django==2.2.20
django-phonenumber-field==5.0.0
djangorestframework==3.12.4
mysqlclient==2.0.3
phonenumbers==8.12.21
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
